% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zissou-package.R
\docType{package}
\name{zissou-package}
\alias{zissou}
\alias{zissou-package}
\title{zissou}
\description{
zissou
}
\examples{
print("hello")
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

